requirements.txt
albumentations
pandas==2.2.2
tqdm==4.66.4
tensorboard==2.18.0
torch==2.6.0
torchvision
transformers
hydra-core==1.3.2
omegaconf==2.3.0
matplotlib
scikit-learn==1.5.2
opencv-python==4.10.0.84
protobuf==3.20.3
torchmetrics==1.6.1
python-Levenshtein==0.25.1
pyarrow==17.0.0
beautifulsoup4
pymorphy2
nltk
pytorch-tabnet